BridgeMenuBar View - replaces code used in bridge list view and bridge photos view
