+++ PittsburghCityBridges/Appearance/PBColorPalate.swift
@@ -11,6 +11,8 @@ import SwiftUI
 struct PBColorPalate {
     var textFgnd: Color = Color(uiColor: UIColor.label)
     var textBgnd: Color = Color(UIColor.secondarySystemBackground)
+    var titleTextFgnd: Color = .pbTitleTextFgnd
+    var titleTextBgnd: Color = .pbTitleTextBgnd
 }
 
 class PBColorPalateSource {
