+++ PittsburghCityBridges/Views/BridgeListView.swift
@@ -25,34 +25,34 @@ struct BridgeListView: View {
     var body: some View {
         NavigationView {
             VStack {
-            menuBar()
-            ScrollView {
-                LazyVStack(spacing: 0, pinnedViews: [.sectionHeaders]) {
-                    ForEach(bridgeListViewModel.sections(groupedBy: bridgeInfoGrouping)) { bridgesSection in
-                        Section {
-                            ForEach(bridgesSection.bridgeModels) { bridgeModel in
-                                NavigationLink(destination: BridgeDetailsView(bridgeModel: bridgeModel, pbColorPalate: bridgesSection.pbColorPalate)) {
-                                    BridgeListRow(bridgeModel: bridgeModel)
-                                        .padding([.leading])
+                BridgeMenuBar(bridgeInfoGrouping: $bridgeInfoGrouping)
+                ScrollView {
+                    LazyVStack(spacing: 0, pinnedViews: [.sectionHeaders]) {
+                        ForEach(bridgeListViewModel.sections(groupedBy: bridgeInfoGrouping)) { bridgesSection in
+                            Section {
+                                ForEach(bridgesSection.bridgeModels) { bridgeModel in
+                                    NavigationLink(destination: BridgeDetailsView(bridgeModel: bridgeModel, pbColorPalate: bridgesSection.pbColorPalate)) {
+                                        BridgeListRow(bridgeModel: bridgeModel)
+                                            .padding([.leading])
+                                    }
+                                    Divider()
                                 }
-                                Divider()
+                                .font(.body)
+                            } header: {
+                                HStack {
+                                    sectionLabel(bridgesSection.sectionName, bridgeInfoGrouping)
+                                        .foregroundColor(bridgesSection.pbColorPalate.textFgnd)
+                                        .font(.title3)
+                                        .padding([.leading])
+                                    Spacer()
+                                }
                             }
-                            .font(.body)
-                        } header: {
-                            HStack {
-                                sectionLabel(bridgesSection.sectionName, bridgeInfoGrouping)
-                                    .foregroundColor(bridgesSection.pbColorPalate.textFgnd)
-                                    .font(.title3)
-                                    .padding([.leading])
-                                Spacer()
-                            }
+                            .font(.headline)
+                            .foregroundColor(bridgesSection.pbColorPalate.textFgnd)
+                            .background(bridgesSection.pbColorPalate.textBgnd)
                         }
-                        .font(.headline)
-                        .foregroundColor(bridgesSection.pbColorPalate.textFgnd)
-                        .background(bridgesSection.pbColorPalate.textBgnd)
                     }
                 }
-            }
             }
             .background(Color.black)
             .navigationBarHidden(true)
@@ -62,7 +62,6 @@ struct BridgeListView: View {
     
     @ViewBuilder
     private func sectionLabel(_ sectionName: String, _ sectionListby: BridgeListViewModel.BridgeInfoGrouping) -> some View {
-        
         switch sectionListby {
         case .neighborhood:
             Text("\(sectionName) Neighborhood")
@@ -70,54 +69,6 @@ struct BridgeListView: View {
             Text("Starting with \(sectionName)")
         case .year:
             Text("Built in \(sectionName)")
-        }
-    }
-    
-}
-
-extension BridgeListView {
-    
-    private func menuBar() -> some View {
-        HStack {
-            Spacer()
-            Text("Pittsburgh Bridges")
-                .foregroundColor(.pbTitleTextFgnd)
-                .font(.title)
-            Spacer()
-            sortMenu()
-                .padding(.trailing, 10)
-        }
-        .background(Color.pbTitleTextBgnd)
-    }
-    
-    private func sortMenu() -> some View {
-        Menu(content: {
-            Button {
-                bridgeInfoGrouping = .name
-            } label: {
-                makeCheckedSortLabel("By Names", selectedSection: .name)
-            }
-            Button {
-                bridgeInfoGrouping = .neighborhood
-            } label: {
-                makeCheckedSortLabel("By Neighborhoods", selectedSection: .neighborhood)
-            }
-            Button {
-                bridgeInfoGrouping = .year
-            } label: {
-                makeCheckedSortLabel("By Year Built", selectedSection: .year)
-            }
-        }, label: {
-            Label("Sort", systemImage: "arrow.up.arrow.down.square")
-                .labelStyle(.titleAndIcon)
-        })
-    }
-    
-    private func makeCheckedSortLabel(_ name: String, selectedSection: BridgeListViewModel.BridgeInfoGrouping) -> Label<Text, Image> {
-        if self.bridgeInfoGrouping == selectedSection {
-            return Label(name, systemImage: "checkmark.square.fill")
-        } else {
-            return Label(name, systemImage: "square")
         }
     }
 }
