+++ PittsburghCityBridges/Views/BridgePhotosView.swift
@@ -26,7 +26,7 @@ struct BridgePhotosView: View {
     var body: some View {
         NavigationView {
             VStack {
-                menuBar()
+                BridgeMenuBar(bridgeInfoGrouping: $bridgeInfoGrouping)
                 ScrollView {
                     LazyVStack(spacing: 20, pinnedViews: [.sectionHeaders]) {
                         ForEach(bridgeListViewModel.sections(groupedBy: bridgeInfoGrouping)) { bridgesSection in
@@ -74,54 +74,6 @@ struct BridgePhotosView: View {
             Text("Starting with \(sectionName)")
         case .year:
             Text("Built in \(sectionName)")
-        }
-    }
-    
-    
-}
-
-extension BridgePhotosView {
-    
-    private func menuBar() -> some View {
-        HStack {
-            Spacer()
-            Text("Pittsburgh Bridges")
-                .foregroundColor(.pbTitleTextFgnd)
-                .font(.title)
-            Spacer()
-            sortMenu()
-        }
-        .background(Color.pbTitleTextBgnd)
-    }
-    
-    private func sortMenu() -> some View {
-        Menu(content: {
-            Button {
-                bridgeInfoGrouping = .name
-            } label: {
-                makeCheckedSortLabel("By Names", selectedSection: .name)
-            }
-            Button {
-                bridgeInfoGrouping = .neighborhood
-            } label: {
-                makeCheckedSortLabel("By Neighborhoods", selectedSection: .neighborhood)
-            }
-            Button {
-                bridgeInfoGrouping = .year
-            } label: {
-                makeCheckedSortLabel("By Year Built", selectedSection: .year)
-            }
-        }, label: {
-            Label("Sort", systemImage: "arrow.up.arrow.down.square")
-                .labelStyle(.titleAndIcon)
-        })
-    }
-    
-    private func makeCheckedSortLabel(_ name: String, selectedSection: BridgeListViewModel.BridgeInfoGrouping) -> Label<Text, Image> {
-        if self.bridgeInfoGrouping == selectedSection {
-            return Label(name, systemImage: "checkmark.square.fill")
-        } else {
-            return Label(name, systemImage: "square")
         }
     }
 }
